Changed types from int/uint32_t to off_t.
The timeout logic was still using int/uint32_t to do the
calculations. Using off_t makes it do the right thing
about connections serving >2GB static files, for example.

+++ src/io.c
@@ -49,7 +49,7 @@ static ssize_t file_recv(IOBuf *iob, char *buffer, int len)
 
 static ssize_t plain_stream_file(IOBuf *iob, int fd, off_t len)
 {
-    ssize_t sent = 0;
+    off_t sent = 0;
     off_t total = 0;
     off_t offset = 0;
     off_t block_size = MAX_SEND_BUFFER;
